Big Additions
1. Done adding rotations to the world altogether from specified camera pin-hole.
2. fixed bug in MeshModel [normals] where it disappears while camera's projection changing.
3. fixed individual MeshModel rotations.
4. Made code clean to each dirty file.

+++ Viewer/include/Trans.h
@@ -8,25 +8,25 @@ public:
 		m[3][3] = 1;
 		return m;
 	}
-	static glm::mat4x4 getZoom(int factor) {
+	static glm::mat4x4 getZoom(float factor) {
 		return glm::mat4x4(glm::highp_mat4::col_type(1.0f, 0.0f, 0.0f, 0.0f),
 			glm::highp_mat4::col_type(0.0f, 1.0f, 0.0f, 0.0f),
 			glm::highp_mat4::col_type(0.0f, 0.0f, factor, 0.0f),
 			glm::highp_mat4::col_type(0.0f, 0.0f, 0.0f, 1.0f));
 	}
-	static glm::mat4x4 getxRotate4x4(int teta) {
+	static glm::mat4x4 getxRotate4x4(float teta) {
 		return glm::mat4x4(glm::highp_mat4::col_type(1.0f, 0.0f, 0.0f, 0.0f),
 			glm::highp_mat4::col_type(0.0f, cos(teta), (-1)*sin(teta), 0.0f),
 			glm::highp_mat4::col_type(0.0f, sin(teta), cos(teta), 0.0f),
 			glm::highp_mat4::col_type(0.0f, 0.0f, 0.0f, 1.0f));
 	}
-	static glm::mat4x4 getyRotate4x4(int teta) {
+	static glm::mat4x4 getyRotate4x4(float teta) {
 		return glm::mat4x4(glm::highp_mat4::col_type(cos(teta), 0.0f, (-1)*sin(teta), 0.0f),
 			glm::highp_mat4::col_type(0.0f, 1.0f, 0.0f, 0.0f),
 			glm::highp_mat4::col_type(sin(teta), 0.0f, cos(teta), 0.0f),
 			glm::highp_mat4::col_type(0.0f, 0.0f, 0.0f, 1.0f));
 	}
-	static glm::mat4x4 getzRotate4x4(int teta) {
+	static glm::mat4x4 getzRotate4x4(float teta) {
 		return glm::mat4x4(glm::highp_mat4::col_type(cos(teta), (-1)*sin(teta), 0.0f, 0.0f),
 			glm::highp_mat4::col_type(sin(teta), cos(teta), 0.0f, 0.0f),
 			glm::highp_mat4::col_type(0.0f, 0.0f, 1.0f, 0.0f),
